src/python/requirements.txt
pillow
streamlit
beautifulsoup4
